+++ src/main/java/com/deccom/domain/core/extractor/ControlVariableExtractorImpl.java
@@ -5,8 +5,8 @@ import com.deccom.domain.core.CVStyle;
 public class ControlVariableExtractorImpl implements ControlVariableExtractor {
 
 	@Override
-	public Integer getData() {
-		return 0;
+	public Double getData() {
+		return 0.0;
 	}
 
 	@Override
